lib/requirements.txt
# =============================================================================
# GBF Story Extractor - Python Dependencies
# =============================================================================
# Install: pip install -r lib/requirements.txt
# Then run: playwright install chromium

# =============================================================================
# Core Dependencies
# =============================================================================

# Web extraction (Playwright for browser automation)
playwright>=1.40.0

# HTTP requests
requests>=2.31.0

# =============================================================================
# Translation APIs
# =============================================================================

# Claude AI translation
anthropic>=0.18.0

# =============================================================================
# Notion Integration
# =============================================================================

notion-client>=2.2.0

# =============================================================================
# CLI & UI
# =============================================================================

# Progress bars
tqdm>=4.66.0

# =============================================================================
# Configuration (Optional)
# =============================================================================

# YAML config support (optional, for config.yaml)
pyyaml>=6.0

# =============================================================================
# Development & Testing
# =============================================================================

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking (optional)
# mypy>=1.5.0

# =============================================================================
# Post-install
# =============================================================================
# After pip install, run:
#   playwright install chromium
